docs: `recover`
